Fixed bug where the first attempt to load a map would fail
When the main scene is added to the tree, only _ready() in main.gd is called before the GameManager continues signalling. This means that maps.gd cannot use the state_changed signals during the transition it is created. I'm gonna try to make sure _ready() is called for everything in the main scene before the GameManager keeps signalling.